refactor: 가구 회전과 카메라 기능(회전, 뷰 모드 전환) 분리
