+++ app/simulator/components/furnitures/FurnitureController.tsx
@@ -3,6 +3,7 @@
 import { useState, useMemo } from 'react';
 
 import { useLightingStore } from '@/stores/useLightingStore';
+import { useViewStore } from '@/stores/useViewStore';
 
 import FurnitureThumbnailInfo from './FurnitureThumbnailInfo';
 import ScaleLockToggle from './ScaleLockToggle';
@@ -16,9 +17,35 @@ import FurnitureControllerBtn from './FurnitureControllerBtn';
 import { useFurnitureStore } from '@/stores/useFurnitureStore';
 import type { FurnitureStoreInfo } from '@/app/types/furniture';
 
+const VALID_ANGLES = [45, 135, 225, 315];
+const TOP_VIEW_ANGLES = [0, 90, 180, 270];
+
+function convertForm3DToTop(angle3D: number): number {
+  const index3D = VALID_ANGLES.indexOf(angle3D);
+  if (index3D !== -1) {
+    return TOP_VIEW_ANGLES[index3D];
+  }
+  return 0;
+}
+
+function convertFormTopTo3D(angleTop: number): number {
+  const indexTop = TOP_VIEW_ANGLES.indexOf(angleTop);
+  if (indexTop !== -1) {
+    return VALID_ANGLES[indexTop];
+  }
+
+  return 45;
+}
+
 export default function FurnitureController() {
-  const { furnitures, selectedFurnitureId, updateFurniture } = useFurnitureStore();
+  const { furnitures, selectedFurnitureId, updateFurniture } =
+    useFurnitureStore();
   const [isScaleLocked, setIsScaleLocked] = useState(false);
+
+  const angle = useViewStore((s) => s.angle);
+  const setAngle = useViewStore((s) => s.setAngle);
+  const isTopView = useViewStore((s) => s.isTopView);
+  const setIsTopView = useViewStore((s) => s.setIsTopView);
 
   const isDay = useLightingStore((state) => state.isDay);
   const setIsDay = useLightingStore((state) => state.setIsDay);
@@ -33,6 +60,28 @@ export default function FurnitureController() {
     }
   };
 
+  // 방 회전
+  const handleRoomRotate = () => {
+    const currentAngles = isTopView ? TOP_VIEW_ANGLES : VALID_ANGLES;
+    const currentIndex = currentAngles.indexOf(angle);
+    const nextIndex = (currentIndex + 1) % currentAngles.length;
+    const newAngle = currentAngles[nextIndex];
+    setAngle(newAngle);
+  };
+
+  // 뷰 모드 전환
+  const handleViewToggle = () => {
+    if (isTopView) {
+      const new3DAngle = convertFormTopTo3D(angle);
+      setAngle(new3DAngle);
+      setIsTopView(false);
+    } else {
+      const newTopAngle = convertForm3DToTop(angle);
+      setAngle(newTopAngle);
+      setIsTopView(true);
+    }
+  };
+
   return (
     <div className='w-[219px] px-[30px] py-[25px] rounded-[30px] bg-gradient-to-r from-white/10 to-black/20 backdrop-blur-md shadow-[0_0_15px_#00000026] flex flex-col items-center justify-center gap-4'>
       {/* 카메라 모드 버튼 */}
@@ -41,8 +90,12 @@ export default function FurnitureController() {
           width={14}
           height={14}
           icon='/assets/icons/view-mode.svg'
+          onClick={handleRoomRotate}
         />
-        <FurnitureControllerBtn text='3D 뷰' />
+        <FurnitureControllerBtn
+          text={isTopView ? '3D 뷰' : '2D 뷰'}
+          onClick={handleViewToggle}
+        />
       </div>
 
       {/* 조명 모드 버튼 */}
