+++ app/simulator/components/furnitures/FurnitureControllerBtn.tsx
@@ -2,27 +2,6 @@
 
 import { useEffect, useState } from 'react';
 import Image from 'next/image';
-import { useViewStore } from '@/stores/useViewStore';
-
-const VALID_ANGLES = [45, 135, 225, 315];
-const TOP_VIEW_ANGLES = [0, 90, 180, 270];
-
-function convertForm3DToTop(angle3D: number): number {
-  const index3D = VALID_ANGLES.indexOf(angle3D);
-  if (index3D !== -1) {
-    return TOP_VIEW_ANGLES[index3D];
-  }
-  return 0;
-}
-
-function convertFormTopTo3D(angleTop: number): number {
-  const indexTop = TOP_VIEW_ANGLES.indexOf(angleTop);
-  if (indexTop !== -1) {
-    return VALID_ANGLES[indexTop];
-  }
-
-  return 45;
-}
 
 interface FurnitureControllerBtnProps {
   icon?: string;
@@ -39,43 +18,16 @@ export default function FurnitureControllerBtn({
   height,
   onClick,
 }: FurnitureControllerBtnProps) {
-  const angle = useViewStore((s) => s.angle);
-  const setAngle = useViewStore((s) => s.setAngle);
-  const isTopView = useViewStore((s) => s.isTopView);
-  const setIsTopView = useViewStore((s) => s.setIsTopView);
-
   const [viewMode, setViewMode] = useState(text);
 
-  const isViewToggleButton = text === '3D 뷰' || text === '2D 뷰';
-  const isRotateRoomButton = icon && !text;
-
   useEffect(() => {
-    if (isViewToggleButton) {
-      setViewMode(isTopView ? '2D 뷰' : '3D 뷰');
-    }
-  }, [isTopView, isViewToggleButton]);
+    setViewMode(text);
+  }, [text]);
 
   const handleClick = () => {
-    if (isViewToggleButton) {
-      if (isTopView) {
-        const new3DAngle = convertFormTopTo3D(angle);
-        setAngle(new3DAngle);
-        setIsTopView(false);
-        setViewMode('3D 뷰');
-      } else {
-        const newTopAngle = convertForm3DToTop(angle);
-        setAngle(newTopAngle);
-        setIsTopView(true);
-        setViewMode('2D 뷰');
-      }
-    }
-
-    if (isRotateRoomButton) {
-      const currentAngles = isTopView ? TOP_VIEW_ANGLES : VALID_ANGLES;
-      const currentIndex = currentAngles.indexOf(angle);
-      const nextIndex = (currentIndex + 1) % currentAngles.length;
-      const newAngle = currentAngles[nextIndex];
-      setAngle(newAngle);
+    if (onClick) {
+      onClick();
+      return;
     }
   };
